Naming cleanup:  - Move the 'core' classes from 'mysql' Java package to 'core' package.  - Change 'mysql-socket-factory' to 'jdbc-socket-factory' in non-mysql    specific artifact IDs.
