+++ connector-j-5/pom.xml
@@ -4,7 +4,7 @@
 
     <parent>
         <groupId>com.google.cloud.sql</groupId>
-        <artifactId>mysql-socket-factory-parent</artifactId>
+        <artifactId>jdbc-socket-factory-parent</artifactId>
         <version>1.0.4-SNAPSHOT</version>
     </parent>
     <artifactId>mysql-socket-factory</artifactId>
@@ -25,7 +25,7 @@
         </dependency>
         <dependency>
             <groupId>com.google.cloud.sql</groupId>
-            <artifactId>mysql-socket-factory-core</artifactId>
+            <artifactId>jdbc-socket-factory-core</artifactId>
             <version>1.0.4-SNAPSHOT</version>
         </dependency>
     </dependencies>
