+++ connector-j-6/src/main/java/com/google/cloud/sql/mysql/SocketFactory.java
@@ -16,6 +16,7 @@
 
 package com.google.cloud.sql.mysql;
 
+import com.google.cloud.sql.core.SslSocketFactory;
 import java.io.IOException;
 import java.net.Socket;
 import java.util.Properties;
