+++ core/pom.xml
@@ -4,10 +4,10 @@
 
     <parent>
         <groupId>com.google.cloud.sql</groupId>
-        <artifactId>mysql-socket-factory-parent</artifactId>
+        <artifactId>jdbc-socket-factory-parent</artifactId>
         <version>1.0.4-SNAPSHOT</version>
     </parent>
-    <artifactId>mysql-socket-factory-core</artifactId>
+    <artifactId>jdbc-socket-factory-core</artifactId>
     <packaging>jar</packaging>
 
     <name>Cloud SQL MySQL Socket Factory (Core Library, don't depend on this directly)</name>
